medical-ragbot/requirements.txt
# ========================================
# MediVault RAG Bot — Requirements
# ========================================
# System prerequisites (NOT pip-installable):
#   Tesseract OCR: https://github.com/UB-Mannheim/tesseract/wiki
#   Poppler:       https://github.com/oschwartz10612/poppler-windows/releases
# ========================================

# ========== PDF Processing ==========
pdfplumber>=0.11.0
pdf2image>=1.17.0
pypdf>=5.0.0

# ========== OCR ==========
pytesseract>=0.3.10
Pillow>=11.0.0

# ========== Vector Store & Database ==========
pymongo>=4.6.0

# ========== LLM & Embeddings ==========
# Local embeddings via ONNX — no PyTorch needed, ~50MB RAM on Render free tier
fastembed>=0.3.0

# HTTP client for Groq / Ollama API
requests>=2.31.0
groq>=1.0.0

# OpenAI SDK (used in llm_handler for API-compatible calls to Groq)
openai>=1.10.0

# ========== API Framework ==========
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# ========== Data Validation ==========
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ========== Utilities ==========
python-dotenv>=1.0.0
numpy>=1.26.0
tenacity>=8.2.0
tqdm>=4.66.0

# ========== LangChain (retrieval helpers) ==========
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0

# ========== Testing (dev only — not needed in production) ==========
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# ========== PDF Generation (test PDFs) ==========
reportlab>=4.0.0
